+++ src/ens.ts
@@ -1,5 +1,5 @@
 import { default as ensInterface } from './ens/contract/ens';
-import { default as resolverInterface } from './ens/contract/resolver';
+import { default as resolverInterface, OldResolverAddresses } from './ens/contract/resolver';
 import { default as hash, childhash } from './ens/namehash';
 import { formatsByCoinType } from '@ensdomains/address-encoder';
 import {
@@ -237,10 +237,10 @@ export default class Ens extends EthereumNamingService {
     return this.ensureRecordPresence(domain, key, record);
   }
 
-  private async getResolverContract(domain: string): Promise<Contract> {
+  private async getResolverContract(domain: string, coinType?: number): Promise<Contract> {
     const resolverAddress = await this.resolver(domain);
     return this.buildContract(
-      resolverInterface(resolverAddress),
+      resolverInterface(resolverAddress, coinType),
       resolverAddress,
     );
   }
@@ -301,7 +301,6 @@ export default class Ens extends EthereumNamingService {
     ]);
   }
 
-  /** @internal */
   protected getCoinType(currencyTicker: string): number {
     const constants: Bip44Constants[] = require('bip44-constants');
     const coin = constants.findIndex(
@@ -316,8 +315,8 @@ export default class Ens extends EthereumNamingService {
     return coin;
   }
 
-  private async fetchAddressOrThrow(resolver, nodeHash, coinType: number) {
-    if (!resolver || isNullAddress(resolver)) {
+  private async fetchAddressOrThrow(resolver: string, nodeHash, coinType: number) {
+    if (isNullAddress(resolver)) {
       return null;
     }
     const resolverContract = this.buildContract(
@@ -325,15 +324,15 @@ export default class Ens extends EthereumNamingService {
       resolver,
     );
     const addr: string =
-      coinType != EthCoinIndex
+      coinType !== EthCoinIndex
         ? await this.callMethod(resolverContract, 'addr', [nodeHash, coinType])
         : await this.callMethod(resolverContract, 'addr', [nodeHash]);
-    if (!addr || addr === '0x') return null;
+    if (isNullAddress(addr)) return null;
     const data = Buffer.from(addr.replace('0x', ''), 'hex');
     return formatsByCoinType[coinType].encoder(data);
   }
 
-  private async fetchAddress(resolver, nodeHash, coin = EthCoinIndex) {
+  private async fetchAddress(resolver, nodeHash, coin) {
     return (
       (await this.ignoreResolutionError(
         ResolutionErrorCode.RecordNotFound,